[CodeGenLib] Add validation step for input in IDLToStructuredSwiftTranslator
Motivation:

The IDLToStructuredSwiftTranslator should discover errors in the CodeGenerationRequest object, before the specialized translators start their work. This way, if the input is not correct no translator will be started.

Modifications:

Moved the validation functions into a IDLToStructuredSwiftTranslator extension and the associated tests in a separate test file.

Result:

The input validation is not done by each individual specialzed translator, but by the main translator, before transformations take place.
