+++ Tests/GRPCCodeGenTests/Internal/Translator/TypealiasTranslatorSnippetBasedTests.swift
@@ -497,131 +497,6 @@ final class TypealiasTranslatorSnippetBasedTests: XCTestCase {
       server: true
     )
   }
-
-  func testTypealiasTranslatorSameNameServicesNoNamespaceError() throws {
-    let serviceA = ServiceDescriptor(
-      documentation: "Documentation for AService",
-      name: "AService",
-      namespace: "",
-      methods: []
-    )
-
-    let codeGenerationRequest = makeCodeGenerationRequest(services: [serviceA, serviceA])
-    let translator = TypealiasTranslator(client: true, server: true)
-    XCTAssertThrowsError(
-      ofType: CodeGenError.self,
-      try translator.translate(from: codeGenerationRequest)
-    ) {
-      error in
-      XCTAssertEqual(
-        error as CodeGenError,
-        CodeGenError(
-          code: .nonUniqueServiceName,
-          message: """
-            Services in an empty namespace must have unique names. \
-            AService is used as a name for multiple services without namespaces.
-            """
-        )
-      )
-    }
-  }
-
-  func testTypealiasTranslatorSameNameServicesSameNamespaceError() throws {
-    let serviceA = ServiceDescriptor(
-      documentation: "Documentation for AService",
-      name: "AService",
-      namespace: "namespacea",
-      methods: []
-    )
-
-    let codeGenerationRequest = makeCodeGenerationRequest(services: [serviceA, serviceA])
-    let translator = TypealiasTranslator(client: true, server: true)
-    XCTAssertThrowsError(
-      ofType: CodeGenError.self,
-      try translator.translate(from: codeGenerationRequest)
-    ) {
-      error in
-      XCTAssertEqual(
-        error as CodeGenError,
-        CodeGenError(
-          code: .nonUniqueServiceName,
-          message: """
-            Services within the same namespace must have unique names. \
-            AService is used as a name for multiple services in the namespacea namespace.
-            """
-        )
-      )
-    }
-  }
-
-  func testTypealiasTranslatorSameNameMethodsSameServiceError() throws {
-    let methodA = MethodDescriptor(
-      documentation: "Documentation for MethodA",
-      name: "MethodA",
-      isInputStreaming: false,
-      isOutputStreaming: false,
-      inputType: "NamespaceA_ServiceARequest",
-      outputType: "NamespaceA_ServiceAResponse"
-    )
-    let service = ServiceDescriptor(
-      documentation: "Documentation for AService",
-      name: "AService",
-      namespace: "namespacea",
-      methods: [methodA, methodA]
-    )
-
-    let codeGenerationRequest = makeCodeGenerationRequest(services: [service])
-    let translator = TypealiasTranslator(client: true, server: true)
-    XCTAssertThrowsError(
-      ofType: CodeGenError.self,
-      try translator.translate(from: codeGenerationRequest)
-    ) {
-      error in
-      XCTAssertEqual(
-        error as CodeGenError,
-        CodeGenError(
-          code: .nonUniqueMethodName,
-          message: """
-            Methods of a service must have unique names. \
-            MethodA is used as a name for multiple methods of the AService service.
-            """
-        )
-      )
-    }
-  }
-
-  func testTypealiasTranslatorSameNameNoNamespaceServiceAndNamespaceError() throws {
-    let serviceA = ServiceDescriptor(
-      documentation: "Documentation for SameName service with no namespace",
-      name: "SameName",
-      namespace: "",
-      methods: []
-    )
-    let serviceB = ServiceDescriptor(
-      documentation: "Documentation for BService",
-      name: "BService",
-      namespace: "SameName",
-      methods: []
-    )
-    let codeGenerationRequest = makeCodeGenerationRequest(services: [serviceA, serviceB])
-    let translator = TypealiasTranslator(client: true, server: true)
-    XCTAssertThrowsError(
-      ofType: CodeGenError.self,
-      try translator.translate(from: codeGenerationRequest)
-    ) {
-      error in
-      XCTAssertEqual(
-        error as CodeGenError,
-        CodeGenError(
-          code: .nonUniqueServiceName,
-          message: """
-            Services with no namespace must not have the same names as the namespaces. \
-            SameName is used as a name for a service with no namespace and a namespace.
-            """
-        )
-      )
-    }
-  }
 }
 
 extension TypealiasTranslatorSnippetBasedTests {
